% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumOp.R
\name{enumOp}
\alias{enumOp}
\title{Enumerated list of operations and their symbols}
\usage{
enumOp()
}
\description{
Enumerated list of operations and their symbols
}
\examples{
if (getRversion() > "4.0") {
  sd(cars$speed)
  sd(cars$dist)
}

}
