% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotter}
\alias{plotter}
\title{Plot cars data}
\usage{
plotter()
}
\description{
Plot cars data
}
\examples{
if (require("ggplot2")) {
 ggplot(cars, aes(x = speed)) + geom_histogram(bins = 20)
 ggplot(cars, aes(x = speed, y = dist)) + geom_point()
}
}
