% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdCars.R
\name{std}
\alias{std}
\title{Calculate Standard Deviation}
\usage{
std(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Calculate Standard Deviation
}
\examples{
if (require("stats")) {
  std(cars$speed)
  std(cars$dist)
}
}
