% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdCars.R
\name{stdCars}
\alias{stdCars}
\title{Standard deviations for cars data}
\usage{
stdCars()
}
\description{
Standard deviations for cars data
}
\examples{
if (require("stats")) {
 sd(cars$speed)
 sd(cars$dist)
}
}
