% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumOp.R
\name{enumOp}
\alias{enumOp}
\title{Enumerated list of operations and their symbols}
\usage{
enumOp()
}
\description{
Enumerated list of operations and their symbols
}
\examples{
enumOp()

library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) + geom_line(size = 0.5)
}
