% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotter1}
\alias{plotter1}
\alias{plotter2}
\title{Plot cars data}
\usage{
plotter1()

plotter2()
}
\description{
Plot cars data
}
\examples{
if (require("ggplot2")) {
 ggplot(cars, aes(x = speed)) + geom_histogram(bins = 20)
 ggplot(cars, aes(x = speed, y = dist)) + geom_point()
}

\donttest{
if (require("ggplot2")) {
  ggplot(cars, aes(x = speed)) + geom_histogram(bins = 20)
  ggplot(cars, aes(x = speed, y = dist)) + geom_point()
}
}

}
