% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdCars.R
\name{std}
\alias{std}
\title{Calculate Standard Deviation (SD)}
\usage{
std(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Calculate Standard Deviation (SD)
}
\examples{
\dontshow{if (getRversion() > "4.0") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
sd(cars$speed)
sd(cars$dist)
\dontshow{\}) # examplesIf}
}
